requirements.txt
Flask
Flask-Cors
python-dotenv
langchain
langchain-community
transformers
torch
sentence-transformers
